pyproject.toml
[project]
name = "lucid-mcp"
version = "0.1.0"
description = "Lucid MCP server built on graphiti-core"
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "mcp>=1.9.4",
    "openai>=1.91.0",
    "graphiti-core[falkordb]>=0.28.2",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
providers = [
    "google-genai>=1.62.0",
    "anthropic>=0.49.0",
    "groq>=0.2.0",
    "voyageai>=0.2.3",
    "sentence-transformers>=2.0.0",
]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.404",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.10"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
